requirements.txt
langchain
langchain-openai
langchain-community
langchain-astradb
datasets
pypdf
python-dotenv
flask
-e .